Add Watch Mode as an npm script & change Mode
Added the Watch Mode npm script to instruct webpack to "watch" all files
for any changes (if they are in the dependency graph). This makes it so
we don't have to run the build manually every time there's a change.
Switched over to development mode from default production in the
webpack config.