Add boilerplate HTML, foundation CSS & comments
Added the boilerplate to the template HTML in src as a starting point.
Added some of the basic foundation CSS elements to the root and body of
the HTML. Added a comment regarding not needing the title in the
plugins section of the webpack config if the title is in the boilerplate
(possibly).