Add function to play audio on logo click
Wrote the function to play the fart audio whenever the user clicks the
app logo next to the title header on the page. Wrote a comment for
implementing other ways to interact with the logo for accessibility
purposes (keydown, etc). Updated comments in the CSS for potential
logo sizing in the future (CSS selector change). Wrapped the logo image
HTML element inside it's own div to target it better in the CSS and JS.