Import audio and image files to JS
Imported the audio and image files to set up relative file path from the
HTML document correctly. Edited the src links in the img elements in
the HTML thinking that was the issue, but the fact that the files
themselves weren't being imported to the JS was the actual issue. Made
some additional comments regarding alternate ways to import the files
and next steps resolving errors and finding an audio/sound loader for
webpack. Moved the audio file to it's own sounds folder to resolve a
separate error in DevTools.