Modify audio play function to resolve all errors
Modified the audio play function due to the audio not playing and
throwing an error message instead any time the user clicked on the
image logo on the webpage. Instead of having the audio element already
exist in the HTML, I changed the function to have the JS dynamically
generate the audio element as a child under the logo image and wired up
the audio file to the generated element. Made some modifications to
existing DOM elements, removed other DOM elements and commented out the
existing audio HTML element to get this to work. Also changed the
parent container of the image from a div to a button in the HTML.
Changed the image filepath from relative to absolute to reduce JS code.
It works but there is a slight delay in the audio play due to the file
itself, NOT the code. Will try to resolve this in the future.