Remove comments no longer needed
Removed all TODO comments regarding next steps as they have been
completed up to this point. Removed commented out audio element in HTML
as JS now generates the element dynamically. Edited the reset timer for
the audio playback on the logo if a user multiclicks. Now resets to the
1 second mark which is about where the audio actually starts (much
closer to what we want). Still an initial delay on the first click
however (not sure how to solve this other than get another file).

Basically the JS and HTML files are cleaned up of unnecessary comments.