Implement first dialog form with JS functions
Added all necessary elements for first dialog modal form in the HTML.
This one is responsible for adding the project itself. The tasks are
planned to be implemented in a sub-dialog form nested under this first
one. The tasks will be added to the project using that form, then the
project will be added to the main area. Created the JS file for the
dialog form and coded the first few classes and functions for the form
to work.