Add button functionality & import/export elements
Exported and imported all elements (variables, functions, classes, etc)
from projects.js to index.js files as a set up for pushing added tasks
to the main area in the future. Added button functionality to the Add
Task button that will be in the heading of the app as well as button
functionality to the Confirm and Cancel buttons in the dialog pop-up
form. Created an animated transition for the dialog form (will
experiment with a different transition in the future).