Remove blur effect transitions & add next steps
Removed all instances of any blur effect transitions in the CSS and
chose to stick with the opacity transitions for the dialog popup
instead. Added some dev comments both in the HTML and projects.js
regarding next steps on how the checkbox feature should be implemented
(on the project itself; haven't gotten to the tasks yet).