Add styling to heading and logo & refactor html
Added some styling similar to the "Holy Grail" layout to the entire
desktop app, mainly the heading section. Resized the logo image button.
Refactored all dialog elements with "project" in the name to "task" as
the form will be repurposed to it's original intent: adding tasks.
Adding projects will be dealt with separately. Gave the form dialog
some styling (using Library project as a template for now).