Create and add code logic to JS file for Projects
Created the separate JS file to be used for the logic/code needed for
the "Add Project" feature in the future. Added some of the code that
was similar to adding tasks into the projects.js file. Renamed the
"title" DOM elements and parameters in the tasks.js file to better
reflect the title in that code being for adding titles to Tasks, not
just creating titles in general. Named the titles in the Projects JS
file in a similar way. Revised some dev comments and deleted the task
argument/parameter for the Task class as well as any other references
to it in tasks.js file.