Rename tags for Task form buttons
After creating the "Add Project" form, I realized that I couldn't keep
the exact same naming convention for the "Confirm" and "Cancel" buttons
on both forms. Renamed the Task form buttons and changed all id tags
everywhere to reflect the new names. Did the same for the buttons in
the "Add Project" form to make them unique as well.