Add more rendered elements to taskDisplay function
Added more elements to be rendered in the taskDisplay function once the
"Add Task" form is submitted. More elements still need to be added and
appended to the right HTML elements (as well as a conditional for
whether or not there is an existing project ul element). Added the
"checked" and "id" number parameters to the Task Class as a way for the
JS to identify which tasks need to be deleted later on and to track
whether or not a task has been completed (checkbox). Made some changes
to comments in the projects.js file.