Append task elements under Title for UI testing
Appended the Date, Time (temporary parameter), Priority level and
Description under the Name of the task. The purpose of this is to
see how all the details regarding the task will look under an <li>
element in the UI once more updates are made to the CSS. Changed the
Project Name from an h2 element to the ul element. The goal is to
append all tasks under the project ul and show them in a "list" format.
Imported the projectDisplay to the tasks JS in order to append the Task
Name under the Project Name. Made some changes to some dev comments and
reorganized some other comments to make the code less confusing/more
readable.