Reorder and refactor classes and "add" functions
Moved all class objects as well as any functions that had to do with
adding a project or task to the index.js file in an attempt to better
organize the code and run more smoothly at runtime. Refactored all the
class objects back to object constructors to prevent any coupling
issues in the future. Also refactored the functions responsible for
adding any tasks or projects to arrays in an attempt to resolve any
previous errors that were popping up on form submit. Will test the
changes in the future. The constructors may have to be moved to their
own separate JS file to keep code as clean and readable as possible.