Modify querySelectorAll and classList naming
Changed all the naming for all classList and querySelectorAll elements
inside projects.js in order to address errors that were popping up when
the projects would try to render to the page. Exported the projects
array out of index.js and imported it into both the tasks and projects
JS files to address an undefined error (both files are able to see
the array now).