Fix bug preventing project from generating
Fixed all errors that prevented the project from being added to the
sidebar as a button and to the main menu as text. Issue now is that
when another project is added, a copy of the previous project is also
added. Likely an issue where a splice method is needed on the array.