Modify project button logic to resolve errors
Modified the child that was being appended to the main area on any
existing project button click to the <ul> DOM instead of the template
literal that tried to represent the actual text entered in the form.
Added a splice method at the end of the display function to remove the
project that was made from the array to prevent duplication errors when
more projects were added. Added a TODO comment in index.js file as a
reminder for adding the splice method there as well if needed.