Modify comments in index and projects files
Added example code for tabbed browsing when switching between projects
in the sidebar via buttons in the projects.js file. Added a reminder at
bottom of index.js file to append whatever function is going to be
responsible for showing the 'default' view where all tasks will be
shown/added when the user opens the webpage.