Refactor forEach button method to index.js
Moved the forEach method for the sidebar project buttons out of the
projectDisplay method in project.js to the index.js file as a function
by itself. This was to try and resolve the error of the buttons removing
the wrong children (always replaces previous projects and only shows
the last one made no matter which project button is clicked). This was
also to organize the code to show with the rest of the button logic
that's in index.js. Created project buttons now don't work at all when
clicked.