Add new modules & refactor code base
Added the new modules to place the render functions for the UI into as
well as for adding the localStorage feature with its own functions in
the future. Refactored nearly the entire JS code base. All application
logic has been refactored. This is anything to do with creating,
storing, removing, updating and toggling the status of tasks and
projects. Deleted any old constructor functions responsible for making
tasks or projects as they are now factory functions. Commented in some
testing code in case it's needed later on (initial tests in another IDE
went well). Renamed some of the existing functions as some of the names
were misleading as far as what the functions actually did.