Refactor appState object to resolve RefError
Changed the parameter of the addProject property in the appState
function from "myProject" back to "projectTitle". This removed the
reference error of "projectTitle not defined" that was showing up
further down in the object (the findProject function). However, the
project simply doesn't render at all even though the confirm button
shows a valid entry in the form and closes the form afterwards. Removed
all comments having to do with "myProject" as a parameter. Renamed the
"Default" project to "Today" as we want the default project to be one
that the user has to do "today" when they open the app. Revised some
description comments in both the projects and index files to better
reflect what certain functions do. Added some more comments regarding
the current error and why findProject method still may not work if used.