Move code to load default project to index
The test code comment I had in projects file that would load the default
project on page load was moved from the projects file to it's proper
place in the index file. It is still commented out as UI functions may
need to be wired up to it in order to get it to work (and still have to
resolve the "blank project" load error).