Modify code in addProject method for debugging
Attempted to debug the error of the project not being added to its
respective array and not being rendered by the renderProject function
for the UI. Created project is not showing up as a sidebar button
anymore or even showing up in the myProjects array currently. Everything
else in the Confirm button logic seems to work (even the dialog closes
afterwards like it added the project to the UI). Made some additional
comments, but not sure on next steps for debugging.