Modify code for method that adds tasks to projects
Tested out different methods in the storeTask function to try and get
the task from the HTML form to be added to an existing project. Getting
TypeErrors with every attempt. Have now realized that there is no option
in the form to add the task to any project or even the default "Today"
one. Changed and added some new TODO comments throughout the code in
all modified files. Imported functions from tasks file into the projects
file for use in functions.

Next steps include replacing "Time" in the Task form with an option to
add the Task to an existing project or default to the "Today" project
including changing the Time option in the HTML form element to whatever
we're using for adding the task to a project (likely a drop down menu).
Then we may need to code all of that in the JS. Also need to get the
default "Today" project to show in the UI on page load before even
modifying the Task form.