Modify function call in storeTask and comments
Revised comments related to test functions. Revised comments related
to parameter for storeTask. Attempted to debug task not being added to
existing project by combining two separate function calls into a single
call, but to no avail. Commented the attempt as well as the error
received (it's the same TypeError as the others).