Add func call to render default project on open
Added the necessary code to get the default project to render on page
load. We took the route to call the existing renderProject function
instead of writing a separate function just for rendering the default
project by itself. This does mean that the default project gets it's
own sidebar button etc. This approach may cause issues in the future
however (button functionality when trying to update or delete the
default project for example). Added some comments regarding current
errors when trying to add the task as well as possibly needing to
rewrite the renderTask function from the ground up as there's a lot of
nesting of objects and arrays going on. Removed comments no longer
needed.