Remove all Time variables and modify Date method
Removed all variables in the renderTask function related to the Time
section of the Task form as it will be replaced with some type of "Add
to Project" section.

Added a toString method next to the Date method for the id variable in
the createTask function.

Removed comments throughout several functions in the tasks file that
were no longer needed.