Replace Time section with Add Project in Task form
Replaced the Time section in the Task HTML form with an Add to Project
section. Gave the Add to Project section a "dynamic" drop-down menu
where the goal is for the user to select from projects they have created
to add their task to. The default "Today" project will be the one the
task gets added to if no other projects exist or if no other project is
selected.

Wrote a function that populates all the projects in the
myProjects array within the drop-down menu for the Add to Project
section. Called that function to all "Add Task" buttons.

Currently the function does not quite work like it should:
either it shows what the console log is, that being [Object object] or
a TypeError stating that 'projectTitle' can't be read/is undefined shows
up and nothing is shown in the drop-down menu.

Modified some comments regarding the possible removal of the current
"Today" button in the top Sidebar section of the HTML as a "Today"
button for the default project auto generates on page load now.