Modify populateProjects function to fix bug
Refactored populateProjects to run on a "for...in" loop similar to the
other render functions in the UI file. Projects now seem to populate
correctly in the drop-down menu of the Add Task form. Removed comments
that were no longer necessary and added a comment regarding the issue
of array index duplication showing up again in the renderProjects
function since splice has been removed (again to prevent hiccups in
the future like not being able to update or remove existing projects).