Comment in coding logic for allowing project dupes
Commented in the coding logic that would allow projects with duplicate
names to be made inside the createProject function, though it may need
to go under appState instead. May implement this feature in the future.

Made some additional comments regarding use of splice method in the
renderProject function in the UI file.