Delete hardcoded HTML Today button in main menu
Deleted the hardcoded Today HTML button that was in the main menu of
the sidebar as now a Today button renders under the My Projects section
when the default Today project renders on page load. The hardcoded
button is no longer needed. Deleted a DOM element related to that
hardcoded Today button. Modified some comments throughout the
project.