Modify renderProject with conditional statement
Modified renderProject slightly by placing all code under a conditional
statement that was supposed to run only if the project in the array
doesn't match the user input in the project form. This was in an
attempt to stop the function from duplicating projects in the array
that were already rendered prior, but it has no effect.

Declared the project's name again under renderTask in order to get the
Task form to submit the user input info in the UI. It works but none of
the info put into the form shows in the UI. Instead all text shows as
'undefined'. The elements themselves do render correctly (in their
proper places).