Modified renderProject conditional statement
Changed the conditional statement of the renderProject function by
adding a 'continue' flow statement in an attempt to get the for...in
loop to stop rendering duplicate existing projects every time a new
project was made. It didn't work (and stops the default project from
rendering on page load). Will do a bit more research but everything
seems to point to refactoring away from for...in loop entirely.