Modify renderProject in attempt to debug errors
Tried several different solutions that were commented inside of the
renderProject function earlier to try and resolve the duplicate
project rendering error while keeping the 'for...in' loop. None of
them worked and the default project stops rendering on page load. The
only other option to keep the for...in loop would maybe be to implement
some type of 'find()' method on the array itself but will probably not
go this route and completely refactor the function from the ground up
instead.