Refactor entire code base of projects.js file
Refactored the appState entirely from an object to a factory function.
Moved all the project methods responsible for creating, updating,
finding and deleting projects outside of the appState object into the
global scope in order to create a separation of concerns. appState is
now only responsible for holding the projects array itself, a function
that gets a project the user will target in the UI, and for creating
the default project which will still be rendered on page load. The
module looks much closer in uniform to the tasks module.

Moved the addTask functionality over to the tasks module under the
storeTask function to clean up the code further.