Clean up imports, exports and organize other code
Revised the imports and exports from projects module to index module.
Gave the tasks array from the projects module global scoping in the
tasks module so all the task functions have access to the array. Moved
the appState function in the projects module to the top of the module
to prevent any errors regarding certain variables not being declared
yet. Added a new variable that retrieves the entire project array
separately in order to keep the actual project array private in the
appState function.