Modify renderTask function in UI module
Modified the for...in loop for the renderTask function to test out in
the future once we figure out how to get the task to render to the
project selected by the user in the UI form. Changed the naming of
the DOM that fetches the <ul> to prevent DRY. Imported tasks array from
the tasks module in order for the new for...in loop to "work" (it still
might not work).

Added the same tasks declaration to the import list on the index module.
Made a comment regarding the renderProject function on the current
for...in loop (basically made a copy of it) since we will be changing
to another type of loop for debugging purposes.