Modify code to test out new loop for renderProject
Modified the for...in loop inside of renderProject to a while loop in
order to fix the duplicate rendering issue. Needed to export a
variable that I forgot to do in an earlier update to get it to run.
While loop doesn't seem to be working. The same TypeError of not being
able to read an undefined property as all other solutions so far,
except this time its for trying to read 'push'. This error occurs
any time I try to do a different loop outside of for...in loop (and
may occur now regardless after the appState refactoring).