Revert appState back to previous version
Removed the factory function version of appState and reverted back to
the object literal. Removed all the functions except createProject and
placed them back in under the appState object like before. Added the
toggle project function into the appState object as well. Renamed the
findProject method to readProject. Reverted the exports to the previous
version.