Revised error message and previous code
Revised an error message in the removeTask function. Commented in a
description of the tasks variable at the top of the tasks module to
show exactly where it's coming from (making it even more obvious).
Corrected some variables, parameters and general naming in some old
code that has been commented out but is saved just in case in both
modules.