Revise storeTask in a TypeError debug attempt
Revised the storeTask function in the tasks module in an attempt to
debug an "Uncaught TypeError: Cannot read properties of undefined
(reading 'push')". Tried adding 'tasks' variable as a parameter to the
storeTask function but get the same error. Likely an issue with the
way I'm calling something either in the 'tasks' variable itself or how
I'm doing it in the storeTask function.

Revised some comments regarding what's happening with the project
button sidebar logic after a new project is created and rendered.