Revise comments on previous loop code
Attempted to retry the loop code from before after getting the app
back to it's previous state, before the refactoring that broke
everything. The 'while' loop and the 'for' loop that is supposed to
read only the last index/item in the array still don't work, but now
they throw different errors. Reverted back to 'for...in' loop for now
until we can come up with another solution.

Added a comment at the top of the projects module as a reminder to try
and use the project's ID as a parameter/target instead of the project's
Title for all possible functions and methods within that module.