Fix the duplicate project rendering bug
Modified the renderProject function in the UI module to take a simple
conditional based on the form data instead of a loop that went through
the nested objects in the array itself. Due to this, I also needed to
create a separate function of renderDefault that would solely be
responsible for rendering the initial default "Today" project on page
load since that's the only project that doesn't need to be rendered via
page form. Modified the function call on the index module to reflect
this change.

The app no longer renders duplicate projects that were previously
rendered!!

Attempted to solve the issue of the renderTask function
not working properly by changing the parameter in the 'for...in' loop
thinking it wasn't targeting the 'tasks' nested array correctly. This
did not work however.

Attempted to resolve the TypeError showing 'push' method reading as
undefined in the storeTask function inside the tasks module by changing
the notation of where the task was being pushed but the same error
persists.