Download and import Update and Delete icons
Found two free icons to use for updating projects and tasks that were
rendered as well as deleting them. Downloaded and imported both icon
png files into the project's img file system. Updated footer of HTML
with proper credits to the artist. These icons will be rendered in
along with all the other elements of any task or project created by
the user.