Add Update/Delete button DOM & Update logic
Added Update/Delete button DOM methods into index module. Also started
Update button logic with comments on 3 different approaches to the
code within. We will VERY likely go with Option 2. Because of this, I
added some form element code as a comment in the HTML file along with
another note about possibly making the drop-down menu for the Add to
Project section of the Task form a little bigger in size.