Add button logic for the Update button on projects
Modified the HTML to include separate forms to update the existing
projects and forms. Tasks is still commented out for now as we are
still working on finishing the logic for Project updates. Added some of
the button logic for updating Projects to the index module. Made
necessary changes for creating, exporting and importing new DOM
elements to make sure the Update buttons work properly in the UI module.

We ended up going with Option #2 to set things up (per comments in the
updateProjectBtn logic in the index module).