Modify button functionality for task update form
Wrote the button functionality code necessary for the Update button in
the Task Update form to work correctly. Refactored the Cancel Task
button to work the same way when the Task Update form is showing as
well (no conflicts between forms). Partially wrote the button
functionality for the Delete Project button that shows next to each
Project on the UI main area. This is not complete yet however (only
takes the project off the array, not the UI).