Add Update & Delete project button functionality
Added the functionality for the Delete button on each project. Added
some functionality to the Update button on each project as well,
specifically the Update button in the Update Project form. This button
is still not working correctly as it currently removes everything from
the project, though there is no code stating to delete or replace any
children, just to change the text of the 'ul' element itself.

Refactored the updateProject method under the appState object in the
Projects module. This was an attempt to resolve the issue with the
Update Project form button, but to no avail. So I doubt the method is
the issue. Although it did resolve a separate TypeError.

Made some comments in the Update Project form button functionality
regarding possible next steps in the HTML file itself.