Refactor update/edit Project code to correct bugs
Refactored the updateProject method in the projects module and the edit
Project button logic in the index module to get the UI to update the
Project title via update form correctly, which it now does. However,
the original Projects array is not updating with the new title (still
shows the old one). Updated the id tag for the project update form in
the HTML as good practice. Updated all DOM elements pointing to that id.
Added in a container/wrapper to put all the project elements in for the
render functions related to any project in the UI module.