Fix type and reference errors and remove comments
Fixed error regarding how populateProjects was being called in the Add
Task buttons. Needed to be updated to reflect how the UI module is now
being imported into the index module. Fixed an error in the delete
button logic where it tried to delete the project via an old element.
The UI now creates a wrapper div to place all the project elements in,
so now the wrapper itself will be the target to delete. Removed some
commented out notes and code in the UI module that are no longer needed.