Modify updateProject method trying to debug
Made several changes to the updateProject method under the appState
object in an attempt to debug the issue where the UI would update to
the new name of a project, but the object inside the myProjects array
itself would not. Nothing has worked so far. Confirmed that the
readProject method is working correctly however.