Refactor update button logic and "tags"
Changed the "tags" being rendered on the Update buttons in the UI module
from 'id' to 'class' for better targeting in the future and to prevent
footguns. This was in an attempt to solve the issue of the Update
buttons on subsequent projects created not working, but the "Tags"
weren't the issue. The issue seems to be Event Delegation (need to add
the event listener to the wrapper/container, NOT the buttons themselves
in order for the right projects to update). Started to implement this in
the index module, but it's not done yet - need to do separation of
concerns with the buttons that bring up the Update form and need to
refactor the Update Confirmation button inside the form itself.