Refactor "Update" form button for Project Updates
Refactored the "Update Project" form button logic for submitting title
updates for existing projects in the UI. The reason why is because
there is still a persisting issue where the form will only update the
first project in the main area no matter what other project update
button was used to change another project. It was hitting the very
first <p> tag in the <main> container always. Have not quite finished
the logic but it will likely work via a forEach method and somehow
figuring out another way to represent the current text in the <p> tag
when updating it. It needs to target the right <p> tag and show that
as the current text, not just what's in the first <p> tag of the <main>
area.