Modify Update button and Update form button logic
Attempted to refactor both the Update Project buttons in each project
showing in the UI as well as the Update button in the Update Project
dialog form. The purpose was to try and get both the correct sidebar
project button and the correct project title to update. But now none of
the Update Project buttons work again (as previously). Will try a
completely different (more standard) approach towards how the projects
should be rendering going forward. Instead of showing all of them at
once, will show them in the UI one at a time, which will require
refactoring the code again (both in the UI and index modules).