Modify Update Project confirm button logic
Made one final attempt to fix the bug of the confirm button in the
Update Project form always updating the first button and title in the
main area instead of the one related to the Update button that was
clicked inside the project itself. All buttons function again except now
ALL project sidebar buttons and ALL project titles in the main area of
the UI will update to the value that was typed into the form.