Refactor Update Project icon buttons to fix bugs
Refactored the Update Project icon button functions and the Update form
button functions and was finally able to fix the error that was causing
all sidebar button and project title text to update on any project that
was selected in the main area once the form was submitted. The main
issues were slightly modifying the event targeting of the buttons and
making sure the current text of the title of the project was captured
by the JS before the form even popped up so the right button and title
would be updated on submission. The 'Update Project' feature officially
works now!

Deleted some comments that were no longer needed and moved some old
code that was commented out to the bottom of each module for possible
reference in the future since the project still isn't done.