Refactor 'Delete' button functionality on projects
Refactored the button logic for all the 'Delete Project' icon buttons
for each project in the main area. This was in an attempt to resolve
bugs mentioned in the previous commit. The first project would be
deleted from the array as well as the sidebar button of the first
project, regardless of the project targeted by the user. The sidebar
button error has been corrected and the project in the main area of
the UI itself also deletes correctly. The wrong project being removed
in the array is still an issue.

Attempted to move the NodeList variable delcaration for all the sidebar
project buttons from the Update Button functionality to the global DOM
to prevent DRY as the Delete Button logic uses the same variable for
targeting which sidebar button to delete. This just caused an error
where the button would no longer update or delete correctly.