Refactor deleteProject method in project module
Refactored the deleteProject method inside of the appState object in
the projects module due to it not updating the array correctly, despite
showing the correct project targeted for deletion by the user. I
mistakenly thought the splice method to delete an item from an array
could be used to target the element/item itself, but found out it must
target the index of the item/element. Otherwise it won't work properly,
which is what was happening with my implementation of the method. This
has been corrected and now the delete button features on each project
in the UI work properly (UI and array update correctly now).

Added a TODO comment in the index module regarding possibly fixing form
submission via Enter key on a keyboard. Right now the Enter key is set
to the Cancel button. This is possibly due to how I coded Cancel to be
the "default" highlight button in the HTML, so it may require me to
change it...though it's probably better if I don't if that's the case.