Reinstate search bar function in the app header
Decided to reinstate the search bar that was in the project in a prior
update but was removed in an attempt to get an MVP as quickly as
possible. Gave the search bar the same basic CSS styling as before.
Added in some JS functionality to filter out projects (and hopefully
tasks in the future) when searching for them.

Adding the search bar back in coincides with the app UI now showing
multiple projects in the main area going forward instead of just one at
a time. So a search bar makes a little more sense again.