Refactor CSS to get closer to planned UI design
Needed to refactor most of the CSS Grid elements in order to get the
app to look more like the UI layout that was planned near the beginning
of the project. It is closer but not quite there yet. Gave the footer
some basic styling to match the Grid changes (and so I could see it
again as it would disappear before the styling was added).

Added some notes to the HTML and the index module concerning the search
button that was added in the beginning. We will use it as a backup if
the auto filter on the search bar doesn't work correctly. Otherwise it
will be deleted in the future.

Removed some comments and code that was no longer needed throughout all
files.