Modify element positioning and add comments
Refactored some of the CSS to position some elements closer to intended
UI design. Placed the image logo button and app name text in a row
instead of a column and centered the text. "Centered" the search bar
but it's a temporary fix (margin won't get it done since the bar will
not stay centered as projects are added). Managed to move the main area
towards the right to the right edge of the window, which is what we
want.

Removed the 'e' TODO comment for the poopSound function for the image
logo button as the button works without referencing the event itself.
Added a comment to optionally add the Search button in the index module
referencing the HTML if the auto filter function doesn't work when we
test it out later on.