Comment in alternate grid setup for container
Commented in some CSS that represents an alternative grid setup. This
was initially going to be for the main 'content' wrapper, which is why
it's in that selector currently. However, I realized it would work
better in the 'container' wrapper since that actually has the sidebar
and main content. At least some form of it would work better. Made a
comment noting this.

Added the same styling in the 'left-heading' to the 'center-heading' to
see if that is a better way to center the search bar instead of the
'margin' selector targeting the search bar itself. Will test this out
later on.