Refactor styling of multiple elements
Resized the grid for the main wrapper of the app which contains the
header, another container with the sidebar and main sections, and the
footer. Centered the search bar in the header but it's still not where
we need it to be. Wrote a comment regarding this as well as finding a
cleaner way to lengthen the search bar (other than 'width') as well as
making the bar itself thicker. Resized the sidebar to take up less
space in the second wrapper section so the main area could take up more
space. Centered the credits text in the footer.