Add sidebar button styling with transition effects
Added color, sizing, a transition effect and other styling elements to
all the sidebar buttons. Set the sizing to be the same regardless of
text entered for now. May change how the sizing is determined in the
future depending on how long the text entered in a form can be.

Tested out search bar functionality and it works! Deleted all comments
regarding adding a search button as a backup as it's no longer needed.