Modify and add developer comments and TODOs
Modified existing developer comments in the tasks module to provide
more clarity on next steps after assessing the current state of the
app logic in the module. Added some more TODO items as well. Modified
the comments regarding the toggle method in the projects module as it's
becoming more likely that toggling projects as complete or not will be
optional or in a future branch, not in the MVP.

Will visit the task functions in the UI module to assess the situation
there next, followed by the task button logic in the index module.