Refactor storeTask function to fix push() errors
Refactored the storeTask function inside the tasks module as it wasn't
adding the task entered in the UI form to the correct project object
inside the myProjects array (or any object/array for that matter). The
function now adds the task to the correct project. Added some comments
regarding next steps and additional clarity in both the tasks and UI
modules.