Refactor renderTask function and add readTask
Wrote a readTask function in the tasks module for use in updating and
deleting tasks from project objects later on. Refactored the renderTask
function in the UI module to render tasks based on the values entered
in the tasks form instead of trying to get the values directly from the
array. Keeping a separation of concerns between app logic and UI. This
required changing the 'for...in' loop to a conditional that checks for
each required form section to be filled out correctly.

Changed the targeting of the 'isChecked' variable based on a prior TODO
comment. Commented out the data-attribute that was being added to the
'taskName' variable for the time being. It may not be needed at all.
Will test above changes to see how the UI is being rendered in a future
update.