Refactor renderTask function to fix render errors
Added a function that finds the correct project to add the rendered
task to in the UI as a TypeError was showing up due to the renderTask
function trying to find a <ul> element that didn't exist. The
renderProject function creates the project's title under a <p> element.
Used the same 'projectNameHeader' variable declaration afterwards as
before but instead of targeting the DOM directly, it now is equal to the
new function with the "Add to Project" value as a parameter. The tasks
now render correctly under the right project, but the description is in
the wrong order due to where the conditional statement is placed (the
description elements render and append under it, before the rest of the
elements do). Will reorganize in a future update.

Switched the taskFind variable inside the updateTask function to just
use the readTask function instead. Before it literally returned the same
code as what was in readTask. Added some comments regarding possibly
using findIndex for everything as an option instead.